Former-commit-id: 6a99bfb6124848f9859d885ca9f3e86f18fcf1f6

+++ ConsoleApplication/GeometricSearchModule/src/org/janelia/it/workstation/gui/geometric_search/viewer/gl/oitarr/ArrayCubeGLActor.java
@@ -25,14 +25,14 @@ public class ArrayCubeGLActor extends GL4SimpleActor
     public IntBuffer vertexBufferId=IntBuffer.allocate(1);
 
     List<Vector4> voxels;
-    int xSize;
-    int ySize;
-    int zSize;
+    float xSize;
+    float ySize;
+    float zSize;
     float getVoxelUnitSize;
 
     public Vector4 color=new Vector4(0.0f, 0.0f, 0.0f, 0.0f);
 
-    public ArrayCubeGLActor(List<Vector4> voxels, int xSize, int ySize, int zSize, float voxelUnitSize) {
+    public ArrayCubeGLActor(List<Vector4> voxels, float xSize, float ySize, float zSize, float voxelUnitSize) {
         this.voxels=voxels;
         this.xSize=xSize;
         this.ySize=ySize;
